+++ discord/cogs/persistmessages/__init__.py
@@ -1,0 +1,5 @@
+from .persistmessages import PersistMessages
+
+
+async def setup(bot):
+    await bot.add_cog(PersistMessages(bot))